server/Backend/requirements.txt
Django<3
djongo==1.2.38
sqlparse==0.2.4
dnspython==2.0.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-cors-headers==3.7.0
django-gmailapi-backend==0.2.0
Django-Verify-Email==0.0.5
django-spa==0.3.5
whitenoise==5.0.1